move liquidations tab to deposit views
